+++ app/Http/Controllers/CriteriaController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\TMDB;
+use App\Services\TmdbClient;
 use App\Services\MovieService;
 
 class CriteriaController extends Controller
 {
-    public function index(TMDB $tmdb, MovieService $ms)
+    public function index(TmdbClient $tmdb, MovieService $ms)
     {
         session()->forget('userInput');
