+++ app/Http/Controllers/HomeController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\TMDB;
+use App\Services\TmdbClient;
 
 class HomeController extends Controller
 {
-    public function index(TMDB $tmdb)
+    public function index(TmdbClient $tmdb)
     {
         return view('home', ['trending' => $tmdb->trending()]);
     }
